Point where bot works!
At this point bot does:
It all starts with the fact that the VIP enters the bot,
sees the description. Next, he is invited to start the
interaction with the start command, after which the bot
asks to enter the name, regular key and reward size.
This data is stored in the database. Now the VIP can
forward the message, thereby rewarding the specified
reward size to the user whose message he forwarded.
In addition, the VIP can re-enter his data, including
the regular key name and reward size. This is done by
the edit command, which deletes the entry under the
corresponding TG id, after which the bot offers to
refill all the data by running the start command.
Accordingly, one VIP with the given TG id can have
only one entry in the database with the given settings
for the regular key name and reward size.