Make edit_name cmd depend on reg key
Now changing name of the viz_acc leads to
neccessity of changing the reg_key as well.
At the same time a user is allowed to change
reg_key and reward size in regardless of name.
Right now, there is some disregard for DRY principal,
so it's better to consider refactoring the consider
in future, especially for intention of showing in
to potential employers.